+++ src/config.ts
@@ -31,5 +31,6 @@ config.sslKeyPath = config.sslKeyPath && config.sslKeyPath.trim();
 // config.sslCert =
 // config.sslPrivkey =
 config.loudBadTokenErrors = Boolean(config.loudBadTokenErrors);
+config.preload = Boolean(config.preload || true);
 
 export default config as Readonly<AppConfig>;